requirements.txt
aiofiles==0.5.0
arrow==0.17.0
astroid==2.4.2
attrs==20.3.0
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.4.6
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.15
httptools==0.1.1
httpx==0.11.1
Hypercorn==0.11.2
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
isort==5.5.2
lazy-object-proxy==1.4.3
lxml==4.7.1
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.2
pandas==1.1.2
priority==1.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-docx==0.8.11
pytz==2020.1
requests==2.24.0
rethinkdb==2.4.8
rfc3986==1.4.0
sanic==20.6.3
Sanic-Cors==0.10.0.post3
sanic-jwt==1.5.0
Sanic-Plugins-Framework==0.9.4
service-identity==18.1.0
six==1.15.0
sniffio==1.1.0
toml==0.10.1
tzlocal==2.1
urllib3==1.25.10
uvicorn==0.13.4
websockets==8.1
wincertstore==0.2
wrapt==1.12.1
wsproto==1.0.0
zope.interface==5.2.0
